test(pivot): Clean the state of the pivot tests
